% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tibble.R
\name{nest_metadata}
\alias{nest_metadata}
\title{Re-Nest Metadata}
\usage{
nest_metadata(df)
}
\arguments{
\item{df}{A dataframe containing columns to be contained within the metadata column.}
}
\value{
A tibble with a nested metadata column.
}
\description{
This function aggregates file information from multiple folder paths into a single tibble.
}
