% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate Parsed Metadata from Excel Template}
\usage{
validate_metadata(metadata, quiet = FALSE, fail_on_issue = TRUE)
}
\arguments{
\item{metadata}{A list-like object output from `summarize_metadata_from_raw_template()`.}

\item{quiet}{Logical; suppress warning output and printing if TRUE.}

\item{fail_on_issue}{Logical; stop execution if problems are found (default TRUE).}
}
\value{
Returns metadata invisibly if no issues are found, otherwise stops with an error.
}
\description{
Runs a series of checks on extracted metadata to catch issues like unbalanced tags,
whitespace, missing table definitions, or duplicate tags.
}
