% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlr8.R
\name{xlr8_write}
\alias{xlr8_write}
\title{Write Multiple Excel Reports from a List-Tibble}
\usage{
xlr8_write(
  df,
  metadata_path = NULL,
  all_info = NULL,
  wb = NULL,
  sheets = NULL,
  sheets_regex = ".",
  overwrite = FALSE
)
}
\arguments{
\item{df}{A tibble with at least two columns:
- `df`: a list-column of data frames to be written.
- `output_path`: character column with file paths.}

\item{metadata_path}{(Optional) Path to the Excel template to use across all rows.
Ignored if `all_info` or `wb` is supplied.}

\item{all_info}{(Optional) Pre-parsed metadata from `summarize_metadata()`.}

\item{wb}{(Optional) A preloaded workbook object from `wb_load()`.}

\item{sheets}{Optional character vector of sheet names to read when parsing metadata.}

\item{sheets_regex}{Regular expression to match sheet names if `sheets` is NULL.}

\item{overwrite}{Logical. Whether to overwrite existing files. Default is `FALSE`.}
}
\value{
Invisibly returns the input tibble `df`.
}
\description{
Writes multiple Excel workbooks by mapping over a tibble where each row contains
the data and output path. This function wraps `xlr8_write_one()` and is designed
for batch generation of styled Excel files.
}
